Static cast float to uint8 git push

+++ esphome/components/hitachi_ac344/hitachi_ac344.cpp
@@ -180,7 +180,7 @@ void HitachiClimate::transmit_state() {
       break;
   }
 
-  set_temp_(this->target_temperature);
+  set_temp_(static_cast<uint8_t>(this->target_temperature));
 
   switch (this->fan_mode) {
     case climate::CLIMATE_FAN_LOW:
